cast voind to int

+++ 0x0B-malloc_free/1-strdup.c
@@ -50,7 +50,7 @@ int _strlen(char *s)
 int i = 0;
 
 if (s == NULL)
-return (NULL);
+return (int)(NULL);
 while (i >= 0)
 {
 if  (s[i] == 0)
